pyproject.toml
[tool.poetry]
name = "document-hundler"
version = "0.1.0"
description = ""
authors = ["KsyushaKI <oksanaakadem39@gmail.com>"]
readme = "README.md"
packages = [{include = "document_hundler"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.2"
Flask = "^2.3.2"
openpyxl = "^3.1.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
PyPDF2 = "^3.0.1"
flake8 = "^6.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
